feat: move ColPali2 classes to match restructured package
