+++ colpali_engine/compression/__init__.py
@@ -1,0 +1,1 @@
+from .pooling import MultiVectorPooler, MultiVectorPoolingStrategy
