+++ colpali_engine/compression/pooling/__init__.py
@@ -1,0 +1,1 @@
+from .multi_vector_pooler import MultiVectorPooler, MultiVectorPoolingStrategy
