--- colpali_engine/models/late_interaction/colpali_2/colpali_2_modeling_outputs.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-
-import torch
-
-
-@dataclass
-class ColPali2ModelOutput:
-    single_vec_emb: torch.Tensor
-    multi_vec_emb: torch.Tensor
