+++ colpali_engine/models/paligemma/__init__.py
@@ -1,2 +1,8 @@
 from .bipali import BiPali, BiPaliProcessor, BiPaliProj
 from .colpali import ColPali, ColPaliProcessor
+from .colpali_2 import (
+    ColPali2,
+    ColPali2Processor,
+    ColPali2ModelOutput,
+    ColPali2Config,
+)
