+++ colpali_engine/models/paligemma/colpali_2/__init__.py
@@ -1,0 +1,3 @@
+from .configuration_colpali_2 import ColPali2Config
+from .modeling_colpali_2 import ColPali2, ColPali2ModelOutput
+from .processing_colpali_2 import ColPali2Processor
