+++ colpali_engine/models/paligemma/colpali_2/configuration_colpali_2.py
