+++ colpali_engine/models/paligemma/colpali_2/processing_colpali_2.py
